refactor(ir): move IrExtension and ExtendingIrEmitter to extension package
Move the Ir extension API (IrExtension, ExtendingIrEmitter,
applyExtensions) out of the transformer package into a dedicated
community.flock.wirespec.ir.extension package, separating the public
extension entry points from the internal IR transformer machinery.
Update all imports across the integration modules and the Gradle/Maven
plugins.
