+++ src/compiler/ir/src/commonMain/kotlin/community/flock/wirespec/ir/extension/IrExtension.kt
@@ -1,4 +1,4 @@
-package community.flock.wirespec.ir.transformer
+package community.flock.wirespec.ir.extension
 
 import community.flock.wirespec.compiler.core.parse.ast.AST
 import community.flock.wirespec.ir.core.IR
