+++ src/integration/spring/src/jvmTest/kotlin/community/flock/wirespec/integration/spring/java/emit/SpringJavaIrExtensionTest.kt
@@ -15,7 +15,7 @@ import community.flock.wirespec.compiler.utils.noLogger
 import community.flock.wirespec.emitters.java.JavaIrEmitter
 import community.flock.wirespec.integration.spring.emit.SpringMappingAnnotationsSupportExtension
 import community.flock.wirespec.integration.spring.emit.SpringMappingNativeSupportExtension
-import community.flock.wirespec.ir.transformer.applyExtensions
+import community.flock.wirespec.ir.extension.applyExtensions
 import kotlinx.io.buffered
 import kotlinx.io.files.Path
 import kotlinx.io.files.SystemFileSystem
