+++ src/plugin/gradle/src/main/kotlin/BaseWirespecTask.kt
@@ -8,8 +8,8 @@ import community.flock.wirespec.compiler.core.emit.Emitted
 import community.flock.wirespec.compiler.core.emit.Emitter
 import community.flock.wirespec.compiler.core.emit.PackageName
 import community.flock.wirespec.compiler.utils.Logger
-import community.flock.wirespec.ir.transformer.IrExtension
-import community.flock.wirespec.ir.transformer.applyExtensions
+import community.flock.wirespec.ir.extension.IrExtension
+import community.flock.wirespec.ir.extension.applyExtensions
 import community.flock.wirespec.plugin.Language
 import community.flock.wirespec.plugin.io.ClassPath
 import community.flock.wirespec.plugin.io.Directory
